Update 2022-10-11-Welcome-to-the-OGGM-Slack.md
Removing a duplicated label. 

The labels for the blog posts seem to be capital sensitive. While the "community" label is displayed with a capital C the it was in the raw text not. Writing it there with a capital create an additional label that looked precisely the same. This PR will hopefully solve that issue.